Tart architecture support
This adds architecture support to the Tart recipes.

Cirrus Labs began shipping an Intel (x86_64/amd64) build of the app back in January. Previously there was only an Apple Silicon (arm64) release.

The current recipe pulls the Intel release because it is first alphabetically on GitHub, which is likely what users are not expecting.

This change allows admins to specify if they want the "arm64" or "amd64" release.